refactor: rename registration page components
