+++ src/pages/registration/validation-chekcs/Address/ValidatedPostalCodeInput/helpers/isValidPostalCode.ts
