+++ src/pages/registration/validation-chekcs/Address/ValidatedStreetInput/helpers/isValidStreetName.ts
